Add QR Code to item form
