+++ Rsc/Palmkedex_Rsc.rcp
@@ -335,7 +335,8 @@ BEGIN
 	BUTTON "" ID 1501			AT (147 1 10 11) NOFRAME
 	LABEL "Type:" 1503 AT (42 16) FONT 1
 	LABEL "             " 1504 AT (72 17)
-	FIELD ID 1502				AT (0 50 160 160) MAXCHARS 1024 MULTIPLELINES NONEDITABLE HASSCROLLBAR
+	FIELD ID 1502				AT (0 50 160 60) MAXCHARS 1024 MULTIPLELINES HASSCROLLBAR
+	LABEL "Details:" 1505 AT (62 150) FONT 1
 END
 
 #include "Rsc/typesBmps.rcp"
