+++ Src/helpers.c
@@ -7,7 +7,6 @@
 #include "UiResourceIDs.h"
 
 
-
 #ifndef NATIVE_CODE
 
 	#include "myTrg.h"
@@ -68,89 +67,3 @@ void* memcpy (void *d, const void *s, size_t l) {
     return d;
 }
 
-void unregisterCurrentAci(void)
-{
-		struct DrawState *ds;
-
-		ds = (struct DrawState*)globalsSlotVal(GLOBALS_SLOT_DETAIL_ACI_IMAGE);
-
-		if (ds)
-		{
-			imgDrawStateFree(ds);
-			*globalsSlotPtr(GLOBALS_SLOT_DETAIL_ACI_IMAGE) = NULL;
-		}
-}
-
-void FreeDescriptionField(UInt16 objectID)
-{
-		FieldType *fld = GetObjectPtr(objectID);
-		Char *ptr = FldGetTextPtr(fld);
-
-		FldSetTextPtr(fld, (char*)noDexEntryString);
-
-		if (ptr && ptr != (char*)noDexEntryString) {
-			MemPtrFree(ptr);
-		}
-}
-
-void SetItemMainFormTitle(SharedVariables *sharedVars)
-{
-		char titleStr[ITEM_NAME_LEN + 6]; // 6 = space + # + 4nums + null char
-
-		itemNameGet(titleStr, sharedVars->selectedPkmnId);
-		StrCat(titleStr, " #");
-		StrIToA(titleStr + StrLen(titleStr), sharedVars->selectedPkmnId);
-
-		FrmCopyTitle(FrmGetActiveForm(), titleStr);
-}
-
-void DrawItemPlaceholder(Coord x, Coord y)
-{
-		MemHandle h;
-		BitmapPtr bitmapP;
-		h = DmGetResource(bitmapRsc, BmpMissingIcon);
-
-		bitmapP = (BitmapPtr)MemHandleLock(h);
-
-		WinDrawBitmap(bitmapP, x, y);
-		MemPtrUnlock(bitmapP);
-		DmReleaseResource(h);
-}
-
-UInt8 getDanaMode(Coord width, Coord height)
-{
-	if (height == 160 && width > 320)
-		return DANA_LANDSCAPE;
-
-	if (width == 160 && height > 320)
-		return DANA_POTRAIT;
-
-	return 0;
-}
-
-void InnerIterate(WChar c)
-	{
-		SharedVariables *sharedVars = (SharedVariables*)globalsSlotVal(GLOBALS_SLOT_SHARED_VARS);
-
-		UInt16 selected = sharedVars->selectedPkmnId;
-
-		if (c == vchrPageUp)
-		{
-			selected--;
-		}
-		else if (c == vchrPageDown)
-		{
-			selected++;
-		}
-
-		if (selected == 0)
-		{
-			selected = TOTAL_ITEM_COUNT_ZERO_BASED;
-		}
-		else if (selected > TOTAL_ITEM_COUNT_ZERO_BASED)
-		{
-			selected = 1;
-		}
-
-		sharedVars->selectedPkmnId = selected;
-	}
